+++ AggarApi/DATA/DataAccess/Repositories/IRepositories/IVehicleRepository.cs
@@ -10,6 +10,6 @@ namespace DATA.DataAccess.Repositories.IRepositories
 {
     public interface IVehicleRepository : IBaseRepository<Vehicle>
     {
-        public IQueryable<Vehicle> GetNearestVehicles(int? brandId, int? typeId, VehicleTransmission? transmission, string? searchKey, double? minPrice, double? maxPrice, int? year, double? rate);
+        public IQueryable<Vehicle> GetNearestVehicles(int? brandId, int? typeId, VehicleTransmission? transmission, string? searchKey, decimal? minPrice, decimal? maxPrice, int? year, double? rate);
     }
 }
